change loading to less animation
